requirements.txt
openai>=1.14
requests
pynacl
